Update project to version 0.8.0 and add EmptyResponse model
This update bumps the project version from 0.7.1 to 0.8.0, reflected in both pyproject.toml and __init__.py. It also introduces a new model, EmptyResponse, which provides a placeholder response class when a response model is not provided. This improves handling of cases where a response model is None by returning an instance of EmptyResponse instead of raising a ValueError, enhancing error management. Additionally, the version constraint of pydantic has been adjusted to be less than 2, ensuring compatibility.