Update package versions and refactor response handling code
This commit includes an update of versions for the 'attrs' and 'pytest' packages in the `poetry.lock` file. The `pydantic_aiohttp` package has also been bumped from version 1.0.0 to 1.0.1. In addition, the code for handling responses in the `pydantic_aiohttp` module has been refactored for better clarity and separation of different types of responses. Unnecessary import statements have been removed to clean up the code.